chore: improve the directory structure for alert history components
