+++ frontend/src/pages/AlertDetails/AlertHeader/AlertLabels/AlertLabels.tsx
@@ -1,7 +1,7 @@
 import './AlertLabels.styles.scss';
 
-import KeyValueLabel from 'periscope/components/KeyValueLabel/KeyValueLabel';
-import SeeMore from 'periscope/components/SeeMore/SeeMore';
+import KeyValueLabel from 'periscope/components/KeyValueLabel';
+import SeeMore from 'periscope/components/SeeMore';
 
 // eslint-disable-next-line @typescript-eslint/no-explicit-any
 type AlertLabelsProps = {
