+++ frontend/src/periscope/components/DataStateRenderer/index.tsx
@@ -1,0 +1,3 @@
+import DataStateRenderer from './DataStateRenderer';
+
+export default DataStateRenderer;
