+++ frontend/src/periscope/components/KeyValueLabel/index.tsx
@@ -1,0 +1,3 @@
+import KeyValueLabel from './KeyValueLabel';
+
+export default KeyValueLabel;
