+++ frontend/src/periscope/components/SeeMore/index.tsx
@@ -1,0 +1,3 @@
+import SeeMore from './SeeMore';
+
+export default SeeMore;
