+++ frontend/src/periscope/components/Tabs2/index.tsx
@@ -1,0 +1,3 @@
+import Tabs2 from './Tabs2';
+
+export default Tabs2;
